# Added optimizers based on multi tensor apply