--- a/torch/optim/_multi_tensor/adadelta.pyi
+++ b/torch/optim/_multi_tensor/adadelta.pyi
@@ -1,0 +1,5 @@
+from typing import Tuple
+from ..optimizer import _params_t, Optimizer
+
+class Adadelta(Optimizer):
+    def __init__(self, params: _params_t, lr: float=..., rho: float=..., eps: float=..., weight_decay: float=...) -> None: ...