--- a/torch/optim/_multi_tensor/adam.pyi
+++ b/torch/optim/_multi_tensor/adam.pyi
@@ -1,0 +1,5 @@
+from typing import Tuple
+from ..optimizer import _params_t, Optimizer
+
+class Adam(Optimizer):
+    def __init__(self, params: _params_t, lr: float=..., betas: Tuple[float, float]=..., eps: float=..., weight_decay: float=..., amsgrad: bool = ...) -> None: ...