… test/test_optim.py
… torch/optim/_multi_tensor/rmsprop.py
… torch/optim/_multi_tensor/rprop.py
… torch/optim/_multi_tensor/sgd.py