@izdeby 2020-09-24
@ngimel, @vincentqb 
To speed things up, im combining all optimizers to land them much faster once i fix the scalal list pr. 

Few things: 
1) yes, its a big one. if its too big, i can break it up into 2-3 parts? 
2) i've added a [test](https://github.com/pytorch/pytorch/pull/45299/files#diff-d237dd23d765aba7cb30b4c21a5e1a86R299) as discussed on previous iterations
3) Please, see perf results in description.